mlops-azure/models/requirements.txt
# Azure ML and MLflow
azure-ai-ml==1.11.0
azure-identity==1.15.0
azure-storage-blob==12.19.0
mlflow==2.9.2
azureml-mlflow==1.53.0

# ML Libraries
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4

# Model Serving
joblib==1.3.2

# Monitoring
opencensus-ext-azure==1.1.13
applicationinsights==0.11.10

# Database
azure-cosmos==4.5.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
